presentation ready

+++ src/tools/test_mutants.py
@@ -1,51 +1,63 @@
 import compile_and_run
 import json
+import argparse
+import multiprocessing
+import threading
+import os
 
 class MutantTester:
     def __init__(self, mutant_lib, mutants_location, vectors):
         self.mutant_lib = self.load_lib(mutant_lib)
         self.mut_numbers = len(self.mutant_lib["mutants"])
         self.mutant_folder = mutants_location
-        self.vectors = load_lib(vectors)['vectors']
+        self.vectors = self.load_lib(vectors)['vectors']
 
     def load_lib(self, path_to_lib):
         with open(path_to_lib) as json_file:
             data = json.load(json_file)
         return data
 
-    def test_mutant(self, mutant_number):
+    def test_mutant(mutant_number, dict, obj):
         compiler = "gcc";
         options = []
         options.append("-lm")
         mutant = compile_and_run.CVirtualMachine(compiler, options)
-        path_mutant = self.mutant_folder +
+        path_mutant = obj.mutant_folder + \
             "Matrix_inverse_LUP_mut_{}.c".format(mutant_number)
-        mutant.compile(path_mutant, "mutant_{}.exe".format(mutant_number))
+        bool = mutant.compile(path_mutant, "mutant_{}.exe".format(mutant_number))
 
-        for vector in self.vectors:
-            res_mutant = mutant.run("mutant", vector[0])
+        if dict is None:
+            dict = obj.mutant_lib
+
+        if not bool:
+            return False
+
+        for vector in obj.vectors:
+            res_mutant = mutant.run("mutant_{}.exe".format(mutant_number), vector[0])
 
             if vector[1] != res_mutant:
-                this.mutant_lib["mutants"][mutant_number]["vector"] = vector
+                dict["mutants"][mutant_number]["vector"] = vector
                 os.remove('mutant_{}.exe'.format(mutant_number))
-                return true
+                print("Found vector: {}\nKilling mutant: {}".format(vector[1], \
+                    dict["mutants"][mutant_number]))
+                return True
 
-        this.mutant_lib["mutants"][mutant_number]["vector"] = None
+        dict["mutants"][mutant_number]["vector"] = None
         os.remove('mutant_{}.exe'.format(mutant_number))
-        return false
+        return False
 
-    def test_mutants_subset(self, min, max):
+    def test_mutants_subset(min, max, dict, obj):
         for i in range(min, max):
-            bool = test_mutant(i)
+            bool = MutantTester.test_mutant(i, dict, obj)
             if not bool:
-                print("No test vector found for mutant: \n{}\n".format(self.mutant_lib["mutants"][i]))
+                print("No test vector found for mutant: \n{}\n".format(dict["mutants"][i]))
 
-    def test_all_mutants(self):
-        test_mutants_subset(0, self.mut_numbers)
+    def test_all_mutants(self, dict):
+        MutantTester.test_mutants_subset(0, self.mut_numbers, dict, self)
 
     def output_library(self, destination):
         with open(destination, 'w') as outfile:
-            json.dump(self.library, outfile, indent=4)
+            json.dump(self.mutant_lib, outfile, indent=4)
 
 
 def main():
@@ -54,41 +66,43 @@ def main():
     parser.add_argument('library', help='input library')
     parser.add_argument('vectors', help='Vectors to test out')
     parser.add_argument('mutant_location', help='folder where mutated programs are contained')
+    parser.add_argument('destination',  help='destination file for vectors')
+
     parser.add_argument('--mult', default=-1, type=int, help='Split the mutants and generate vectors side by side')
     args = parser.parse_args()
     tester = MutantTester(args.library, args.mutant_location, args.vectors)
     if args.mult == -1:
-        tester.test_all_mutants()
+        try:
+            tester.test_all_mutants(tester.mutant_lib)
+        except KeyboardInterrupt:
+            tester.output_library(args.destination)
+            return
     else:
-        # subset_size = tester.mut_numbers / float(args.mult)
-        # for i in range(tester.mut_number / subset_size):
-        print("inshallah un jour ya multiprocessing")
 
-    # compiler = "gcc";
-    # options = []
-    # options.append("-lm")
-    # sut = compile_and_run.CVirtualMachine(compiler, options)
-    # path_sut = "../sut/Matrix_inverse_LUP.c"
-    # sut.compile(path_sut, "sut")
-    #
-    # mutant = compile_and_run.CVirtualMachine(compiler, options)
-    # path_mutant = "../mutants/Matrix_inverse_mutant.c"
-    # mutant.compile(path_mutant, "mutant")
-    #
-    # f = open("simulation_file.json")
-    # data = json.load(f)
-    #
-    # v = data.get('vectors')
-    # kill_mutant = []
-    #
-    # for i in range(int(len(v) / 10)):
-    #     new_lst = [str(j) for sub in v[i] for j in sub]
-    #     res_sut = sut.run("sut", new_lst)
-    #     res_mutant = mutant.run("mutant", new_lst)
-    #
-    #     if res_sut != res_mutant:
-    #         kill_mutant.append(i)
-    #
-    # print(kill_mutant)
+        manager = multiprocessing.Manager()
+        dict = manager.dict(tester.mutant_lib)
 
-main()
+        threads = []
+        subset_size = int(tester.mut_numbers / args.mult)
+        for i in range(0, tester.mut_numbers, subset_size):
+            endpoint = i + subset_size
+            if endpoint > tester.mut_numbers:
+                endpoint = tester.mut_numbers - 1
+            pro = multiprocessing.Process(target=MutantTester.test_mutants_subset, args=[i, endpoint, dict, tester])
+            threads.append(pro)
+        try:
+            for thread in threads:
+                thread.start()
+            for thread in threads:
+                thread.join()
+        except KeyboardInterrupt:
+            tester.output_library(args.destination)
+            return
+
+        tester.mutant_lib = dict
+        print(dict)
+    tester.output_library(args.destination)
+
+
+if __name__ == '__main__':
+    main()
